requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.62.0
wandb>=0.12.0
tensorboard>=2.8.0
pyyaml>=6.0
scikit-image>=0.19.0
scipy>=1.7.0
lpips>=0.1.4
pytorch-fid>=0.3.0
torchmetrics>=0.11.0
accelerate>=0.20.0
transformers>=4.20.0
einops>=0.6.0
timm>=0.9.0